+++ Kernel/System/DB/mssql.pm
@@ -36,6 +36,11 @@ sub LoadPreferences {
     $Self->{'DB::CaseSensitive'}        = 0;
     $Self->{'DB::LikeEscapeString'}     = '';
 
+    # how to determine server version
+    # @@VERSION returns "Microsoft SQL Server 2012 - 11.0.2218.0 (X64) Jun 12 2012 13:05:25 Copyright..."
+    # we only take what is left of the minus; our version string: "Microsoft SQL Server 2012"
+    $Self->{'DB::Version'} = "SELECT LEFT( @@VERSION, (CHARINDEX ( '-' ,@@VERSION) -2) )";
+
     # dbi attributes
     $Self->{'DB::Attribute'} = {
         LongTruncOk => 1,