+++ Kernel/System/DB/postgresql.pm
@@ -42,6 +42,12 @@ sub LoadPreferences {
     $Self->{'DB::CaseSensitive'}        = 1;
     $Self->{'DB::LikeEscapeString'}     = '';
 
+    # how to determine server version
+    # version string contains a comma, we only need what's on the left of it
+    # example of full string: "PostgreSQL 9.2.4, compiled by Visual C++ build 1600, 64-bit"
+    # our result: "PostgreSQL 9.2.4"
+    $Self->{'DB::Version'} = "SELECT LEFT(VERSION(), (SRTPOS(VERSION(),',') -1))";
+
     # dbi attributes
     $Self->{'DB::Attribute'} = {};